requirements.txt
selenium-4.8.3